reuse dialog in cloud selection pop up
